Rename namespaces from Substrait.Core.Type to Substrait.Core.Types for consistency
