+++ src/Substrait.Core/Types/Compound/Decimal.cs
@@ -1,8 +1,10 @@
-﻿namespace Substrait.Core.Type.Compound;
+﻿namespace Substrait.Core.Types.Compound;
 
-public sealed record IntervalCompound : TypeClass
+public sealed record Decimal : TypeClass
 {
   public required int Precision { get; init; }
+
+  public required int Scale { get; init; }
 
   public override TResult Accept<TResult>(ITypeVisitor<TResult> visitor) => visitor.Visit(this);
 }
