+++ src/Substrait.Core/Types/Compound/FixedBinary.cs
@@ -1,6 +1,6 @@
-﻿namespace Substrait.Core.Type.Compound;
+﻿namespace Substrait.Core.Types.Compound;
 
-public sealed record VarChar : TypeClass
+public sealed record FixedBinary : TypeClass
 {
   public required int Length { get; init; }
