+++ src/Substrait.Core/Types/Compound/FixedChar.cs
@@ -1,4 +1,4 @@
-﻿namespace Substrait.Core.Type.Compound;
+﻿namespace Substrait.Core.Types.Compound;
 
 public sealed record FixedChar : TypeClass
 {
