+++ src/Substrait.Core/Types/Compound/IntervalCompound.cs
@@ -1,6 +1,6 @@
-﻿namespace Substrait.Core.Type.Compound;
+﻿namespace Substrait.Core.Types.Compound;
 
-public sealed record PrecisionTimestamp : TypeClass
+public sealed record IntervalCompound : TypeClass
 {
   public required int Precision { get; init; }
