+++ src/Substrait.Core/Types/Compound/IntervalDay.cs
@@ -1,10 +1,8 @@
-﻿namespace Substrait.Core.Type.Compound;
+﻿namespace Substrait.Core.Types.Compound;
 
-public sealed record Decimal : TypeClass
+public sealed record IntervalDay : TypeClass
 {
   public required int Precision { get; init; }
-
-  public required int Scale { get; init; }
 
   public override TResult Accept<TResult>(ITypeVisitor<TResult> visitor) => visitor.Visit(this);
 }
