+++ src/Substrait.Core/Types/Compound/List.cs
@@ -1,4 +1,4 @@
-﻿namespace Substrait.Core.Type.Compound;
+﻿namespace Substrait.Core.Types.Compound;
 
 public sealed record List : TypeClass
 {
