+++ src/Substrait.Core/Types/Compound/PrecisionTimestampTz.cs
@@ -1,6 +1,6 @@
-﻿namespace Substrait.Core.Type.Compound;
+﻿namespace Substrait.Core.Types.Compound;
 
-public sealed record PrecisionTime : TypeClass
+public sealed record PrecisionTimestampTz : TypeClass
 {
   public required int Precision { get; init; }
