+++ src/Substrait.Core/Types/Compound/Struct.cs
@@ -1,4 +1,4 @@
-﻿namespace Substrait.Core.Type.Compound;
+﻿namespace Substrait.Core.Types.Compound;
 
 public sealed record Struct : TypeClass
 {
