+++ src/Substrait.Core/Types/Simple/Fp64.cs
@@ -1,6 +1,6 @@
-﻿namespace Substrait.Core.Type.Simple;
+﻿namespace Substrait.Core.Types.Simple;
 
-public sealed record I64 : TypeClass
+public sealed record Fp64 : TypeClass
 {
   public override TResult Accept<TResult>(ITypeVisitor<TResult> visitor) => visitor.Visit(this);
 }
