+++ src/Substrait.Core/Types/Simple/I16.cs
@@ -1,6 +1,6 @@
-﻿namespace Substrait.Core.Type.Simple;
+﻿namespace Substrait.Core.Types.Simple;
 
-public sealed record String : TypeClass
+public sealed record I16 : TypeClass
 {
   public override TResult Accept<TResult>(ITypeVisitor<TResult> visitor) => visitor.Visit(this);
 }
