+++ src/Substrait.Core/Types/Simple/String.cs
@@ -1,6 +1,6 @@
-﻿namespace Substrait.Core.Type.Simple;
+﻿namespace Substrait.Core.Types.Simple;
 
-public sealed record Uuid : TypeClass
+public sealed record String : TypeClass
 {
   public override TResult Accept<TResult>(ITypeVisitor<TResult> visitor) => visitor.Visit(this);
 }
